% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_freq_bar}
\alias{plot_freq_bar}
\title{Plot the frequency of a variable on horizontal bars}
\usage{
plot_freq_bar(data, variable)
}
\arguments{
\item{data}{data.frame to use}

\item{variable}{variable we are interested in}
}
\value{
plot of type ggplot
}
\description{
Plot the frequency of a variable on horizontal bars
}
