% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_selectInput.R
\name{set_selectInput}
\alias{set_selectInput}
\title{Create simple selectInput based on a specific variable}
\usage{
set_selectInput(data, variable)
}
\arguments{
\item{data}{input data}

\item{variable}{variable to be filtered on}
}
\value{
a selectInput for the UI of a shiny App
}
\description{
Create simple selectInput based on a specific variable
}
