% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicr-package.R
\docType{package}
\name{musicr-package}
\alias{musicr}
\alias{musicr-package}
\title{musicr: Create a Shiny App for a collection of vinyls and discs}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ludovic Bertelà \email{ludovic.bertela@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
